Unify `IpPacket` types into common library
