+++ rust/connlib/snownet/src/lib.rs
@@ -4,14 +4,12 @@ mod allocation;
 mod backoff;
 mod channel_data;
 mod index;
-mod ip_packet;
 mod node;
 mod ringbuffer;
 mod stats;
 mod stun_binding;
 mod utils;
 
-pub use ip_packet::{IpPacket, MutableIpPacket};
 pub use node::{
     Answer, Client, ClientNode, Credentials, Error, Event, Node, Offer, Server, ServerNode,
     Transmit,
