+++ rust/connlib/tunnel/src/tests.rs
@@ -1,10 +1,11 @@
-use crate::{ip_packet::MutableIpPacket, ClientEvent, ClientState, GatewayState};
+use crate::{ClientEvent, ClientState, GatewayState};
 use connlib_shared::{
     messages::{ResourceDescription, ResourceDescriptionCidr, ResourceId},
     proptest::cidr_resource,
     StaticSecret,
 };
 use ip_network_table::IpNetworkTable;
+use ip_packet::MutableIpPacket;
 use pretty_assertions::assert_eq;
 use proptest::{
     arbitrary::any,
@@ -374,7 +375,7 @@ enum Transition {
 fn icmp_request_packet(source: IpAddr, dst: IpAddr) -> MutableIpPacket<'static> {
     match (source, dst) {
         (IpAddr::V4(src), IpAddr::V4(dst)) => {
-            use pnet_packet::{
+            use ip_packet::{
                 icmp::{
                     echo_request::{IcmpCodes, MutableEchoRequestPacket},
                     IcmpTypes, MutableIcmpPacket,
@@ -394,7 +395,7 @@ fn icmp_request_packet(source: IpAddr, dst: IpAddr) -> MutableIpPacket<'static> 
             ipv4_packet.set_next_level_protocol(IpNextHeaderProtocols::Icmp);
             ipv4_packet.set_source(src);
             ipv4_packet.set_destination(dst);
-            ipv4_packet.set_checksum(pnet_packet::ipv4::checksum(&ipv4_packet.to_immutable()));
+            ipv4_packet.set_checksum(ip_packet::ipv4::checksum(&ipv4_packet.to_immutable()));
 
             let mut icmp_packet = MutableIcmpPacket::new(&mut buf[20..]).unwrap();
             icmp_packet.set_icmp_type(IcmpTypes::EchoRequest);
@@ -405,7 +406,7 @@ fn icmp_request_packet(source: IpAddr, dst: IpAddr) -> MutableIpPacket<'static> 
                 MutableEchoRequestPacket::new(icmp_packet.payload_mut()).unwrap();
             echo_request_packet.set_sequence_number(1);
             echo_request_packet.set_identifier(0);
-            echo_request_packet.set_checksum(pnet_packet::util::checksum(
+            echo_request_packet.set_checksum(ip_packet::util::checksum(
                 echo_request_packet.to_immutable().packet(),
                 2,
             ));
@@ -413,7 +414,7 @@ fn icmp_request_packet(source: IpAddr, dst: IpAddr) -> MutableIpPacket<'static> 
             MutableIpPacket::owned(buf).unwrap()
         }
         (IpAddr::V6(src), IpAddr::V6(dst)) => {
-            use pnet_packet::{
+            use ip_packet::{
                 icmpv6::{
                     echo_request::MutableEchoRequestPacket, Icmpv6Code, Icmpv6Types,
                     MutableIcmpv6Packet,
@@ -445,7 +446,7 @@ fn icmp_request_packet(source: IpAddr, dst: IpAddr) -> MutableIpPacket<'static> 
             echo_request_packet.set_sequence_number(1);
             echo_request_packet.set_checksum(0);
 
-            let checksum = pnet_packet::icmpv6::checksum(&icmp_packet.to_immutable(), &src, &dst);
+            let checksum = ip_packet::icmpv6::checksum(&icmp_packet.to_immutable(), &src, &dst);
             MutableEchoRequestPacket::new(icmp_packet.payload_mut())
                 .unwrap()
                 .set_checksum(checksum);
