+++ rust/snownet-tests/Cargo.toml
@@ -8,6 +8,7 @@ edition = "2021"
 anyhow = "1"
 boringtun = { workspace = true }
 snownet = { workspace = true }
+ip-packet = { workspace = true }
 futures = "0.3"
 hex = "0.4"
 pnet_packet = { version = "0.34" }
